Update Emacs dot file provided by Nic McDonald.
'These are two simple updates. First, having the user use "load" or "load-file" is non-standard. The new "provide" statement at the bottom allows the user to simply "require" the mode. I updated the comment in the header to reflect this. Also, the ".dot" extension was already in the mode but the ".gv" extension wasn't. I added it.'